requirements.txt
playwright==1.44.0
pytest==8.3.5
pytest-playwright==0.5.0
pytest-html==4.1.1
allure-pytest==2.13.5
python-dotenv==1.2.2
faker==25.2.0
requests==2.33.0
pytest-timeout==2.3.1
